[ADVAPP-2868]: refactor - disable-on-archive via Campaign model event
