+++ app-modules/campaign/tests/Tenant/Models/CampaignTest.php
@@ -17,7 +17,7 @@
       in the software, and you may not remove or obscure any functionality in the
       software that is protected by the license key.
     - You may not alter, remove, or obscure any licensing, copyright, or other notices
-      of the licensor in the software. Any use of the licensor’s trademarks is subject
+      of the licensor in the software. Any use of the licensor's trademarks is subject
       to applicable law.
     - Canyon GBS Inc. respects the intellectual property rights of others and expects the
       same in return. Canyon GBS® and Advising App® are registered trademarks of
@@ -34,39 +34,41 @@
 </COPYRIGHT>
 */
 
-namespace AdvisingApp\Campaign\Filament\Actions;
+namespace AdvisingApp\Campaign\Tests\Tenant\Models;
 
 use AdvisingApp\Campaign\Models\Campaign;
-use CanyonGBS\Common\Filament\Actions\ArchiveAction;
-use Filament\Actions\Action;
-use Filament\Notifications\Notification;
-use Throwable;
 
-class ArchiveCampaignAction
-{
-    public static function make(): Action
-    {
-        return ArchiveAction::make()
-            ->label(fn (Campaign $record): string => $record->enabled ? 'Disable and Archive' : 'Archive')
-            ->modalHeading(fn (Campaign $record): string => $record->enabled ? 'Disable and Archive Campaign' : 'Archive Campaign')
-            ->modalSubmitActionLabel(fn (Campaign $record): string => $record->enabled ? 'Disable and Archive' : 'Archive')
-            ->action(function (Campaign $record): void {
-                try {
-                    $record->archive();
+use function Tests\asSuperAdmin;
 
-                    Notification::make()
-                        ->success()
-                        ->title('Campaign archived successfully')
-                        ->send();
-                } catch (Throwable $exception) {
-                    report($exception);
+it('disables an enabled campaign when it is archived', function () {
+    asSuperAdmin();
 
-                    Notification::make()
-                        ->danger()
-                        ->title('Failed to archive campaign')
-                        ->body($exception->getMessage())
-                        ->send();
-                }
-            });
-    }
-}
+    $campaign = Campaign::factory()->enabled()->create();
+
+    // Establish starting state before the side effect.
+    expect($campaign->enabled)->toBeTrue()
+        ->and($campaign->archived_at)->toBeNull();
+
+    $campaign->archive();
+
+    $campaign->refresh();
+
+    expect($campaign->enabled)->toBeFalse()
+        ->and($campaign->archived_at)->not->toBeNull();
+});
+
+it('archives an already disabled campaign without re-enabling it', function () {
+    asSuperAdmin();
+
+    $campaign = Campaign::factory()->disabled()->create();
+
+    expect($campaign->enabled)->toBeFalse()
+        ->and($campaign->archived_at)->toBeNull();
+
+    $campaign->archive();
+
+    $campaign->refresh();
+
+    expect($campaign->enabled)->toBeFalse()
+        ->and($campaign->archived_at)->not->toBeNull();
+});
